Apply fix to Z-axis color filling in 3D graphs.
This applies the patch from Debian Bug #489184 (https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=489184)
against the latest version of GD-Graph. By running the pietest script found in the Debian bug I was able
to confirm this is still a current issue in GD-Graph 1.48, and that this patch still fixes it.