don't align two Y axes if either y[12]_{min,max}_value is defined
https://rt.cpan.org/Public/Bug/Display.html?id=62665

old code unconditionally adds 0 Y point to the Graph with two Y axes
and aligns 0 points on left and right sides.

we don't do it if either one of y1_min_value, y2_min_value,
y1_max_value or y2_max_value is defined.

'two_axes == 2' that was proposed in the ticket is not acceptable as
it's not documented and looks too hackish. If this change fails
drastically and has to be reverted then I would rather go with
"dont_align_y_axis" attribute.